+++ examples/data-sources/iosxe_pim_vrf/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxe_pim_vrf" "example" {
-  vrf = "VRF1"
-}