+++ docs/guides/changelog.md
@@ -10,6 +10,7 @@ description: |-
 ## 0.8.0 (unreleased)
 
 - BREAKING CHANGE: Combine `iosxe_msdp` and `iosxe_msdp_vrf` resources and data sources into single `iosxe_msdp` resource and data source
+- BREAKING CHANGE: Combine `iosxe_pim` and `iosxe_pim_vrf` resources and data sources into single `iosxe_pim` resource and data source
 
 ## 0.7.1